# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateLessonInput:
    properties:
      description:
        type: string
      lesson_date:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - lesson_date
    - title
    type: object
...
  dto.UpdateLessonInput:
    properties:
      description:
        type: string
      lesson_date:
        type: string
      status:
        type: string
      title:
        type: string
    required:
    - lesson_date
    - status
    - title
    type: object
...
info:
  contact: {}
  description: Backend API for Mini-LMS / Teacher Platform
  title: Teacher Platform API
  version: "1.0"
paths:
...
  /courses/{id}/lessons:
    get:
      consumes:
      - application/json
      description: Get all lessons for a course
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get lessons by course
      tags:
      - lessons
    post:
      consumes:
      - application/json
      description: Create a new lesson in a course
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create lesson input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateLessonInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create lesson
      tags:
      - lessons
  /lessons/{id}:
    delete:
      consumes:
      - application/json
      description: Delete lesson by id
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete lesson
      tags:
      - lessons
    get:
      consumes:
      - application/json
      description: Get one lesson by id
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get lesson by id
      tags:
      - lessons
    patch:
      consumes:
      - application/json
      description: Update lesson by id
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update lesson input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateLessonInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update lesson
      tags:
      - lessons
...
  /students/{id}/progress:
    get:
      consumes:
      - application/json
      description: Get aggregated progress for a student
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get student progress
      tags:
      - progress
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"